Migrate FtInputTags to the composition API
