+++ src/renderer/components/ExternalPlayerSettings.vue
@@ -67,7 +67,7 @@ import FtSelect from './ft-select/ft-select.vue'
 import FtInput from './ft-input/ft-input.vue'
 import FtToggleSwitch from './ft-toggle-switch/ft-toggle-switch.vue'
 import FtFlexBox from './ft-flex-box/ft-flex-box.vue'
-import FtInputTags from './ft-input-tags/ft-input-tags.vue'
+import FtInputTags from './FtInputTags/FtInputTags.vue'
 
 import store from '../store/index'
